+++ src/documents/forms/index.js
@@ -1,0 +1,1 @@
+export {default as EditDocumentForm} from './EditDocumentForm';
