+++ src/forms/index.js
@@ -1,4 +1,3 @@
-export {default as EditDocumentForm} from './EditDocument/EditDocumentForm';
 export {default as NewProjectForm} from './NewProjectForm';
 export {default as LinkProjectForm} from './LinkProjectForm';
 export {default as EditProjectForm} from './EditProjectForm';
